refactor: useRemainingTime 훅 추출 및 중복 코드 제거
