+++ src/app/(main)/lesson/[id]/_components/AttendanceFloatingBar/AttendanceFloatingBar.tsx
@@ -1,7 +1,7 @@
 'use client'
 
-import { useEffect, useState } from 'react'
 import Text from '@/components/common/Text'
+import useRemainingTime from '@/hooks/useRemainingTime'
 import type { AttendanceSession } from '@/types/attendance'
 import {
   barStyle,
@@ -27,23 +27,6 @@ interface Props {
   onEnd: () => Promise<void>
 }
 
-function useRemainingTime(expiresAt: string) {
-  const [remaining, setRemaining] = useState('')
-
-  useEffect(() => {
-    const calc = () => {
-      const diff = Math.max(0, new Date(expiresAt).getTime() - Date.now())
-      const m = Math.floor(diff / 60000)
-      const s = Math.floor((diff % 60000) / 1000)
-      setRemaining(`${String(m).padStart(2, '0')}:${String(s).padStart(2, '0')}`)
-    }
-    calc()
-    const id = setInterval(calc, 1000)
-    return () => clearInterval(id)
-  }, [expiresAt])
-
-  return remaining
-}
 
 export default function AttendanceFloatingBar({ session, className, isEnding, onOpenDetail, onEnd }: Props) {
   const remaining = useRemainingTime(session.expires_at)
